feat: cleaned unit test
