+++ packages/mantine/src/components/code-editor/__tests__/CodeEditor.spec.tsx
@@ -1,5 +1,5 @@
 import {loader} from '@monaco-editor/react';
-import {render, screen, waitForElementToBeRemoved} from '@test-utils';
+import {render, screen, userEvent, waitForElementToBeRemoved} from '@test-utils';
 import {useForm} from '../../../form';
 
 import {CodeEditor} from '../CodeEditor';
@@ -13,7 +13,7 @@ describe('CodeEditor', () => {
         vi.clearAllMocks();
     });
 
-    it('renders the monaco editor and a copy to clipboard button', async () => {
+    it('renders the monaco editor, a copy to clipboard button and a search button', async () => {
         render(<CodeEditor label="label" description="description" />);
 
         await waitForElementToBeRemoved(screen.queryByRole('presentation'));
@@ -22,6 +22,7 @@ describe('CodeEditor', () => {
         expect(screen.getByText(/description/)).toBeInTheDocument();
         expect(await screen.findByTestId('monaco-editor')).toBeInTheDocument();
         expect(await screen.findByRole('button', {name: /copy/i})).toBeInTheDocument();
+        expect(await screen.findByRole('button', {name: /search/i})).toBeInTheDocument();
     });
 
     it('shows validation errors underneath the code editor', async () => {
