+++ packages/mantine/src/components/search/Search.tsx
@@ -2,9 +2,6 @@ import {SearchSize16Px} from '@coveord/plasma-react-icons';
 import {ActionIcon} from '@mantine/core';
 
 export interface SearchProps {
-    /**
-     * Called each time the search button is clicked.
-     */
     handleSearch: () => void;
 }
 
@@ -14,4 +11,4 @@ const SearchButton: React.FunctionComponent<SearchProps> = ({handleSearch}) => (
     </ActionIcon>
 );
 
-export const Search: React.FunctionComponent<SearchProps> = ({...others}) => <SearchButton {...others} />;
+export const Search: React.FunctionComponent<SearchProps> = ({...params}) => <SearchButton {...params} />;
